requirements.txt
annotated-types==0.7.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
google-auth==2.38.0
google-genai==1.0.0
idna==3.10
lxml==5.3.1
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyMuPDF==1.25.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
websockets==14.2
xlrd==2.0.1
XlsxWriter==3.2.2
